few more zero-init and 'new py*' updates

+++ src/eve-server/station/Station.h
@@ -98,7 +98,7 @@ public:
     // station methods here for offices, reprocessing, and docking.
     PyRep* GetOffices()                                 { PyIncRef(m_officePyData); return m_officePyData; }  // cached officeData for client call
     int8 GetAvalibleOfficeCount()                       { return maxRentableOffices - m_officeMap.size(); }
-    int64 GetOfficeRentalFee()                          { return m_data.officeRentalFee; }
+    uint32 GetOfficeRentalFee()                         { return m_data.officeRentalFee; }
     void RentOffice(OfficeData& odata);
     uint32 GetOfficeID(uint32 corpID);
     uint32 GetOwnerID()                                 { return m_data.corporationID; }
